dtls.c: Check head on HASH_DELETE
HASH_DEL_PEER() may be called with head == NULL. As this is not
allowed for the uthash macro HASH_DELETE(), an explicit check is
required.

This bug was reported by Achim Kraus <achim.kraus@bosch-si.com> in
ticket #14 '"Stale pointer" to peers'. The suggested fix was moved
to the macro definition in dtls.c to avoid porting it to newer
versions of uthash.h.