debug.[ch]: fix compiler errors when configured as --without-debug
debug.c was originally designed to be present only if tinydtls was
built with debug support, but later modifications did not respect
this intention. To resolve this situation, debug.c is always included
in the resulting library and provides all functions that are not
debug-related. These functions are:

* dtls_package_name()
* dtls_package_version()
* dtls_get_log_level()
* dtls_set_log_level()   (see note below)
* dsrv_log() and its convenience macros

Note: dtls_set_log_level() in non-debug mode restricts the log level
to DTLS_LOG_INFO or lower. When called with higher a value, the log
level is set to DTLS_LOG_INFO instead.