dtls.[hc]: added function dtls_prf_with_current_keyblock()
The new function dtls_prf_with_current_keyblock() takes a session
object and retrieves the keying material that is in effect for
that session. When found, the keyblock is used as key for the
default PRF which then is applied to the specified label and
random seed.

A patch for this function has been proposed by Chul Lee in
ticket #20 "Add function for ownerPSK generation in IoTivity".